+++ packages/core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @fe-free/core
 
+## 4.1.42
+
+### Patch Changes
+
+- feat: add valueType CustomMarkdownModal
+  - @fe-free/icons@4.1.42
+  - @fe-free/tool@4.1.42
+
 ## 4.1.41
 
 ### Patch Changes